code/finetune/NER/src/requirements.txt
torch==1.6.0
transformers==4.1.1
seqeval==0.0.12
pytorch-crf==0.7.2
sentencepiece
tensorboard
numpy
tqdm